+++ CHANGELOG.md
@@ -30,6 +30,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 - Lightning Storm: new status effect (spell) with 24 hour duration, increases lightning bolt damage by 50%, immune to Burning while active, applies Rejuvenation upon expiration, extended by 6 hours in mutual war
 - Burning now increases fireball damage by 100% (was fixed pop growth) and grants immunity to Lightning Storm
 - Burning and Lightning Storm are triggered by 20 successes of the matching spell (in a single hour or up to 30 over a 24 hour period)
+- Damage reduction from SPA/WPA removed
+- Sabotage Boats damage reduced to 2.25% (from 2.5%)
+- Wizards now reduce vulnerability to Fireball and Lightning Bolt by 16% per 0.1 WPA
+- Wizard Guilds now reduce vulnerability to Fireball and Lightning Bolt by 10% per 1% owned (was spy and wizard cost/losses/recovery)
 
 ### Fixed
 - Spell reflect stats are now properly attributed to the dominion that cast the friendly spell